+++ packages/core/src/authoring.ts
@@ -9,6 +9,6 @@
 // Note: `input` is exported in `core.ts` due to:
 // https://docs.google.com/document/d/1RXb1wYwsbJotO1KBgSDsAtKpduGmIHod9ADxuXcAvV4/edit?tab=t.0.
 
-export {InputFunction} from './authoring/input';
-export {InputOptions, InputOptionsWithoutTransform, InputOptionsWithTransform, InputSignal, InputSignalWithTransform, ɵINPUT_SIGNAL_BRAND_WRITE_TYPE} from './authoring/input_signal';
-export {ɵUnwrapDirectiveSignalInputs} from './authoring/input_type_checking';
+export {InputFunction} from './authoring/input/input';
+export {InputOptions, InputOptionsWithoutTransform, InputOptionsWithTransform, InputSignal, InputSignalWithTransform, ɵINPUT_SIGNAL_BRAND_WRITE_TYPE} from './authoring/input/input_signal';
+export {ɵUnwrapDirectiveSignalInputs} from './authoring/input/input_type_checking';
