+++ packages/core/src/authoring/input/input_signal_node.ts
