+++ packages/core/src/core.ts
@@ -16,7 +16,7 @@ export * from './authoring';
 // Input is exported separately as this file is exempted from JSCompiler's
 // conformance requirement for inferred const exports.
 // See: https://docs.google.com/document/d/1RXb1wYwsbJotO1KBgSDsAtKpduGmIHod9ADxuXcAvV4/edit?tab=t.0
-export {input} from './authoring/input';
+export {input} from './authoring/input/input';
 
 export * from './metadata';
 export * from './version';
